add generic AclArrayInput
